+++ src/database/operations/authorization/canCreateInitiativeFieldValue.ts
@@ -1,0 +1,12 @@
+import { generateExistsOperation } from '../generators';
+import type { BaseFieldCategory, Id } from '../../../types';
+
+const canCreateInitiativeFieldValue = generateExistsOperation<{
+	initiativeId: Id;
+	baseFieldCategory: BaseFieldCategory;
+}>('authorization.canCreateInitiativeFieldValue', [
+	'initiativeId',
+	'baseFieldCategory',
+]);
+
+export { canCreateInitiativeFieldValue };